[ADD] Added simple building feature.

+++ Project Customer Group 19/Assets/Scripts/Tile.cs
@@ -12,8 +12,8 @@ public class Tile : MonoBehaviour
     }
 
     //X and Y coord on the map and not in the world.
-    private int x;
-    private int z;
+    private float x;
+    private float z;
     private GameObject leftTile;
     private GameObject rightTile;
     private GameObject topLeftTile;
@@ -25,8 +25,8 @@ public class Tile : MonoBehaviour
     //====================================================
     //               Getter and Setters
     //====================================================
-    public int X { get => x; set => x = value; }
-    public int Z { get => z; set => z = value; }
+    public float X { get => x; set => x = value; }
+    public float Z { get => z; set => z = value; }
     public GameObject LeftTile { get => leftTile; set => leftTile = value; }
     public GameObject RightTile { get => rightTile; set => rightTile = value; }
     public GameObject TopLeftTile { get => topLeftTile; set => topLeftTile = value; }
@@ -35,6 +35,8 @@ public class Tile : MonoBehaviour
     public GameObject BottomRightTile { get => bottomRightTile; set => bottomRightTile = value; }
     public TileType Type { get => type; set => type = value; }
 
+
+    //----------------------DEBUG-------------------------------
     //======================================================
     //              ChangeColor(Color color)
     //======================================================
@@ -54,17 +56,17 @@ public class Tile : MonoBehaviour
         if (leftTile != null)
             existingNeighbours.Add(leftTile);
         if (rightTile != null)
-            existingNeighbours.Add(rightTile);        
+            existingNeighbours.Add(rightTile);
         if (topLeftTile != null)
-            existingNeighbours.Add(topLeftTile);       
+            existingNeighbours.Add(topLeftTile);
         if (topRightTile != null)
-            existingNeighbours.Add(topRightTile);        
+            existingNeighbours.Add(topRightTile);
         if (bottomLeftTile != null)
-            existingNeighbours.Add(bottomLeftTile);        
+            existingNeighbours.Add(bottomLeftTile);
         if (bottomRightTile != null)
             existingNeighbours.Add(bottomRightTile);
 
-        foreach(GameObject tile in existingNeighbours)
+        foreach (GameObject tile in existingNeighbours)
         {
             MeshRenderer mRender = tile.GetComponentInChildren<MeshRenderer>();
             mRender.material.color = color;
